requirements.txt
Django>=5.2
djangorestframework
django-filter
djangorestframework-simplejwt
